+++ app/src/main/AndroidManifest.xml
@@ -13,8 +13,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Views.Register"></activity>
-        <activity android:name=".Views.Login"></activity>
+        <activity android:name=".Views.RegisterView"></activity>
+        <activity android:name=".Views.LoginView"></activity>
         <activity android:name=".Views.ProfileView">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
