+++ app/src/main/java/cat/udl/tidic/a_favour/models/ProfileViewModel.java
@@ -1,19 +1,30 @@
 package cat.udl.tidic.a_favour.models;
 
+import android.content.SharedPreferences;
+import android.util.Base64;
 import android.util.Log;
 import android.widget.ImageView;
+import android.widget.Toast;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
+import com.google.gson.JsonObject;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
 import cat.udl.tidic.a_favour.R;
 import cat.udl.tidic.a_favour.RetrofitClientInstance;
 import cat.udl.tidic.a_favour.User;
 import cat.udl.tidic.a_favour.UserServices;
+import cat.udl.tidic.a_favour.Utils;
+import cat.udl.tidic.a_favour.Views.LoginView;
 import cat.udl.tidic.a_favour.Views.ProfileView;
+import cat.udl.tidic.a_favour.Views.RegisterView;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
@@ -29,7 +40,9 @@ public class ProfileViewModel
 
     public ProfileViewModel()
     {
-        getUser();
+        //getUser();
+        userService = RetrofitClientInstance.
+                getRetrofitInstance().create(UserServices.class);
     }
 
    void getUser()
@@ -64,6 +77,11 @@ public class ProfileViewModel
                //Toast.makeText(ProfileViewModel.this, t.getMessage(), Toast.LENGTH_SHORT).show();
            }
        });
+   }
+
+   public void setUser(String username, String password, SharedPreferences mPreferences){
+
+
    }
 
    public String getUsername()
@@ -126,5 +144,60 @@ public class ProfileViewModel
     {
         Log.d("Profile", "S'ha premut l'opció OPINIONS");
     }
+
+    public void registerUser(String user, String password1, String password2, String email, String phone, RegisterView extra){
+        if (!password1.equals(password2)){ //comprovem que les contrasenyes siguin iguals
+
+            sendMessage("Las contraseñas no coinciden", extra);
+
+        } else if(password1.length() < 5){
+
+            sendMessage("La contraseña tiene que tener mínimo 5 caracteres", extra);
+
+        } else {
+            // Course API requires passwords in sha-256 in passlib format so:
+            String p = password1;
+            String salt = "16";
+            String encode_hash = Utils.encode(p,salt,29000);
+            System.out.println("PASSWORD_ENCRYPTED " + encode_hash);
+
+
+            JsonObject user_json = new JsonObject();
+            user_json.addProperty("username", user);
+            user_json.addProperty("email", email);
+            user_json.addProperty("phone", phone);
+            user_json.addProperty("password", encode_hash);
+
+
+
+            Call<Void> call = userService.registerUser3(user_json);
+            call.enqueue(new Callback<Void>() {
+                @Override
+                public void onResponse(Call<Void> call, Response<Void> response) {
+                    if (response.code() == 200){
+                        sendMessage("Usuario registrado", extra);
+                    }else{
+                        try { //Atrapar error usuari existent / correu existent
+                            sendMessage(Objects.requireNonNull(response.errorBody().string()), extra);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                }
+
+                @Override
+                public void onFailure(Call<Void> call, Throwable t) {
+                    sendMessage("Error", extra);
+                }
+            });
+
+        }
+
+    }
+
+    public void sendMessage(String message, RegisterView extra){
+        Toast.makeText(extra,message, Toast.LENGTH_SHORT).show(); //enviem missatge a la pantalla
+    }
+
 
 }
